refactor(capabilities): :recycle: move code-block + mermaid-diagram to features/mdx, track copy internally
These are MDX-rendering features, not design-system primitives. Moving them out lets
copy-code-button call trackWith directly (features may import lib) — removing the onCopy
prop, the server-side TrackedCodeBlock wrapper, and the RSC server→client function-prop
that broke static prerender. Design-system stays pure.
