--- src/components/features/mdx/tracked-code-block/index.ts
@@ -1,1 +1,0 @@
-export { TrackedCodeBlock } from "./tracked-code-block";
